fix(lint): remove redundant int() wrapper around round() in p2p_secagg_driver
round() with no ndigits returns int in Python 3; the cast was a no-op.
All 16 tests remain passing.

+++ backend/app/infrastructure/security/p2p_secagg_driver.py
@@ -270,7 +270,7 @@ class P2PSecAggDriver:
 
         # Quantize: float → int via fixed-point scaling
         quantized: list[int] = [
-            int(round(w * quantization_scale)) % self._MASK_MODULUS
+            round(w * quantization_scale) % self._MASK_MODULUS
             for w in weights
         ]
         masked = list(quantized)  # copy to apply masks in-place
